Update ModuleService.cfc
the property modules has been added to the ModuleConfig configuration:
		modules = {
			preload = [ 'messageQueue']
		};
this is necessary to load submodules in advance.

for instance if a package of modules is delivered within a module, the containing module might have an interceptor which calls a service to run from a module within. in this case the submodule has to be activated in advance.

the module service has been enhanced with a preloading functionality of this modules.