Update ModuleService.cfc
add capability for registering module packages. since coldbox delivers tons of features in modules and development is done more and more in modules. there is a need for packages of modules to register without a parent module.

All but the packages are registered and nicely grouped in logical structure.

myapplication/
   modules/
      myCoolModule
         modules/
            myCoolSubmodule
            myCoolSubmodule-package/
               myEvenCoolerSubSubModule1
               myEvenCoolerSubSubModule2
      coldbox-package/
         i18n
         cborm
         cbstorages
         javaloader