Add subdomain routing capability
A domain can be specified on a route or with closure.
The incoming domain must match the domain in order for the route to be a match.
Additionally, a wildcard or parameter can exist in the route, like `:username.forgebox.io`.
The value of `username` (in this case) will exist in the `rc` scope.