Add option to define reinit key vs. hard-coded fwreinit
Use cases have arisen for both practical and security reasons that it would be handy to be able to define the name of the variable that would allow the framework to be reinitialized. This change provides the ability to define `ReinitKey` in ColdBox.cfc config to specify that variable name, while maintaining `fwreinit` as the default if it is not defined.