Always register new routes when conditions are provided
Conditions are a way to apply a function check before executing a
route.  However, ColdBox also tries to intelligently merge
routes when only the actions are different.  Conditions are tricky
since they are functions and therefore not equatable or hashable.
Because of this, if a condition is specified on a route always
create a new route.  We will trust the developer not to make duplicate
conditional routes.  And even if they do, ColdBox will handle it just fine.