Dynamically generate queryExecute options + new config to include dsn
This change dynamically builds the options struct which will be passed to queryExecute statements. Additionally, there's a new configuration option 'queryIncludeDsn` which determines whether the dsn should be added to the queryExecute options or not. 

This change is designed to assist ACF engine users workaround the limitation that all queryExecute statements must have an identical dsn/username/passsword when run in a transaction