Update Response Pagination Properties for Case-Sensitive Engines
Depending on what engine you are using, the `Response` object's `pagination` struct will either be in camel case or all in caps.  For example in ACF, it will convert all of the variables `offset`, `maxRows`, etc... to `OFFSET`, `MAXROWS`, etc...   Since the `Response` object could be consumed by JavaScript which is case-sensitive, it's important to deliver consistent output between CFML engines.

This update addresses the problem by explicitly forcing the case of the `pagination` struct to be in camel case.