Update RestHandler - Implement abort logic onAuthenticationFailure
The Method signature had support for the abort, but it did not actually do anything with it.
Add param to javadocs, and used the appropriate code from the onAuthorizationFailure
Also updated the comment to say it returns a 401. 
Confirmed the http status codes struct has the 401 for NOT_AUTHENTICATED not a 403.